Log the message state when throwing an exception.
The monkey is able to generate a situation where we're coming into
ComposeMessageActivity thinking we've got an MMS message, based on
the message flags, but the attachment flags indicate the MMS message
hasn't been initialized. Log the message state so next time this happens,
we can track down the source of the problem.